Added root user check

+++ full-remove.sh
@@ -1,5 +1,11 @@
 echo "# Removing"
 
+if [[ $(id -u) -ne 0 ]]; then
+  echo "Please run as root"
+  exit 1
+fi
+
+
 wg-quick down wg0
 systemctl stop wg-quick@wg0
 systemctl disable wg-quick@wg0
